[engsys] Remove the `openTelemetryCommonJs()` export from dev-tool
It is a workaround for packages that uses older version of core-tracing
libraries. Now it only applies to two packages.  This PR removes the export from
`dev-tool` and have copies of the function in the two communication packages'
test rollup config instead.

Part of #17707
