+++ common/tools/dev-tool/src/config/rollup.base.config.ts
@@ -16,51 +16,6 @@ interface PackageJson {
   devDependencies: Record<string, string>;
 }
 
-/**
- * Gets the proper configuration needed for rollup's commonJS plugin for @opentelemetry/api.
- *
- * NOTE: this manual configuration is only needed because OpenTelemetry uses an
- * __exportStar downleveled helper function to declare its exports which confuses
- * rollup's automatic discovery mechanism.
- *
- * @returns an object reference that can be `...`'d into your cjs() configuration.
- */
-export function openTelemetryCommonJs(): Record<string, string[]> {
-  const namedExports: Record<string, string[]> = {};
-
-  for (const key of ["@opentelemetry/api", "@azure/core-tracing/node_modules/@opentelemetry/api"]) {
-    namedExports[key] = [
-      "SpanKind",
-      "TraceFlags",
-      "getSpan",
-      "setSpan",
-      "SpanStatusCode",
-      "getSpanContext",
-      "setSpanContext",
-    ];
-  }
-
-  const releasedOpenTelemetryVersions = ["0.10.2", "1.0.0-rc.0"];
-
-  for (const version of releasedOpenTelemetryVersions) {
-    namedExports[
-      // working around a limitation in the rollup common.js plugin - it's not able to resolve these modules so the named exports listed above will not get applied. We have to drill down to the actual path.
-      `../../../common/temp/node_modules/.pnpm/@opentelemetry+api@${version}/node_modules/@opentelemetry/api/build/src/index.js`
-    ] = [
-      "SpanKind",
-      "TraceFlags",
-      "getSpan",
-      "setSpan",
-      "StatusCode",
-      "CanonicalCode",
-      "getSpanContext",
-      "setSpanContext",
-    ];
-  }
-
-  return namedExports;
-}
-
 // #region Warning Handler
 
 /**
